Consolidate excelToJson and cpfValidationJson into processValidationJson lambda function
