--- api/src/functions/excelToJson/excelToJson.scenarios.ts
@@ -1,8 +1,0 @@
-import type { ScenarioData } from '@redwoodjs/testing/api'
-
-export const standard = defineScenario({
-  // Define the "fixture" to write into your test database here
-  // See guide: https://redwoodjs.com/docs/testing#scenarios
-})
-
-export type StandardScenario = ScenarioData<unknown>
